added DB library
this got way easier cuz i had a brainwave
and realized that anytime we've successfuly sent any data off the boat we can just delete it. What we can do to avoid complexities later is add timestamps to the filenames so that we will only delete them when confirmed by the server, but in the meantime we start a new file. The boat doesn't need to know what is in the files.